Replace yaml with codf for configuration
As a result, a config file that previously would have looked like this:

    max-requests: 4
    ports:
      - dest: http://USER:PASSWORD@localhost:24339/write?db=DBNAME
        flush-every: 10s
        flush-size: 600000
        max-retries: 7
        read-timeout: 2m
        write-timeout: 2m
        backoff:
          interval: 10s
          grow-by: 1s
          factor: 1.2
        listen:
          # NOTE: udp4:// not supported in old configs
          - '127.0.0.1:24337'
          - '127.0.0.1:24338'

Now looks like this:

    max-requests 4;
    port {
        listen udp4://127.0.0.1:24337 127.0.0.1:24338;
        pass http://USER:PASSWORD@localhost:24339/write?db=DBNAME;
        flush 10s 60000;
        max-retries 7;
        timeout 2m 2m; ' write read
        backoff 10s grow-by 1s factor 1.2;
    }

While it's probably a matter of opinion which one is better, this does
serve as a good test case of codf for me, and gets away from having
YAML as a config format where a typo on a key is silently ignored.

Because of this, there are now stricter checks on the parsed config
file than before.